apps/pwm_test: Fix pin handling for STM32
STM32 PWM driver encodes alternate pin function
at higher bits of pin definition.

Definition of pins with alternate function > 0 are 16 bits long,
where bits 12:15 specify function.

pwm_shell assumed that 8 bits is enough for pin number, but it will
not work for most ST cases.

This changes local variable type to accommodate ST encoding of pins
with alternate function.

+++ apps/pwm_test/src/pwm_shell.c
@@ -182,7 +182,7 @@ static const struct shell_cmd_help start_help = {
 static int
 cmd_start(int argc, char **argv)
 {
-    uint8_t pin = LED_BLINK_PIN;
+    int pin = LED_BLINK_PIN;
     char *tmp;
     char *pwm = "pwm0";
     int freq = 200;
